+++ components/Empty.tsx
@@ -1,0 +1,9 @@
+import styles from "../styles/components/Empty.module.css";
+import React from "react";
+
+// Displayed when there is no data to display
+const Empty: React.FC = () => {
+  return <div className={styles.empty}>There are no Staff.</div>;
+};
+
+export default Empty;
