+++ components/StaffView.tsx
@@ -1,15 +1,6 @@
-import React, { useContext } from "react";
-import { User } from "../models/users";
-import StaffCard from "./StaffCard";
+import React from "react";
 import styles from "../styles/components/LeagueViews.module.css";
-import { APIContext } from "../context/APIContext";
-import Loader from "./Loader";
-import Error from "./Error";
-import useSWR from "swr";
-
-type StaffViewProp = {
-  staff: User[];
-};
+import StaffScroll from "./StaffScroll";
 
 const filters = [
   "Administration",
@@ -26,35 +17,15 @@ const filters = [
   "Level 8",
 ];
 
-const StaffView: React.FC<StaffViewProp> = ({ staff }) => {
-  const client = useContext(APIContext);
-
-  // Use SWR hook to get the data from the backend
-  const { data, error } = useSWR("/api/staff", () => client.getStaff());
-
-  // Dynamically load the data from the backend and render the respective component
-  const staffData: any = (data: User[] | undefined, error: Error) => {
-    if (error) return <Error />;
-    if (!data) return <Loader />;
-
-    // Render the staff
-    return data.map((c: any) => (
-      <StaffCard
-        key={c.id}
-        firstName={c.firstName}
-        lastName={c.lastName}
-        phone_number={c.phoneNumber}
-        email={c.email}
-      />
-    ));
-  };
-
+const StaffView: React.FC = () => {
   return (
     <div className={styles.compContainer}>
       <div className={styles.leftContainer}>
         <h1 className={styles.compTitle}>Staff</h1>
         <div className={styles.compList}>
-          <ul className={styles.scroll}>{staffData(data, error)}</ul>
+          <ul className={styles.scroll}>
+            <StaffScroll />
+          </ul>
         </div>
       </div>
       <span className={styles.spacer} />
