+++ styles/components/Empty.module.css
@@ -1,0 +1,6 @@
+.empty {
+  display: flex;
+  justify-content: center;
+  font-size: 1.5rem;
+  color: #f37121;
+}
